refactor(chapter12): rename case06 files to case06_1 for consistency
