+++ chapter12/src/case06/case06_1.py
